feat(smart-contract): Implement Agent Discovery Oracle with Capability Matching (#248)
