+++ smart-contracts/contracts/agent_registry/src/lib.rs
@@ -33,6 +33,9 @@
 //! all-success means the batch committed; any failure means **no** writes occurred.
 
 mod events;
+pub mod types;
+
+pub use types::*;
 
 use events::{
     AdminChangedEvent, AgentDeregisteredEvent, AgentRegisteredEvent, ErrorReportedEvent,
@@ -40,7 +43,7 @@ use events::{
 };
 use soroban_sdk::{
     contract, contractimpl, contracttype, symbol_short, Address, BytesN, Env, Map, String, Symbol,
-    Val, Vec,
+    TryFromVal, Val, Vec,
 };
 
 #[allow(dead_code)]
@@ -192,6 +195,10 @@ pub enum DataKey {
     /// Ledger number at which the cooldown expires for a deregistering agent.
     /// Key present ⟺ the agent is in the cooldown window.
     BondCooldown(Symbol),
+    /// Cache key for discovery oracle results (temporary storage).
+    DiscoveryCache(DiscoveryQuery),
+    /// Aggregate statistics for the discovery oracle (instance storage).
+    DiscoveryStats,
 }
 
 /// Per-item outcome for batch registration (`Ok(agent_id)` / `Err(code)`).
@@ -697,6 +704,222 @@ impl AgentRegistryContract {
         records
     }
 
+    /// Discover and rank agents matching multi-criteria criteria:
+    /// required capability, max price, min reputation, and max latency.
+    ///
+    /// Returns a list of [`DiscoveryResult`] structs ordered descending by composite score.
+    /// Results are cached in temporary storage for gas efficiency within the same transaction.
+    pub fn discover_agents(env: Env, query: DiscoveryQuery) -> Vec<DiscoveryResult> {
+        let cache_key = DataKey::DiscoveryCache(query.clone());
+
+        // 1. Re-query within same transaction is served free from temporary storage cache
+        if let Some(cached_results) = env
+            .storage()
+            .temporary()
+            .get::<_, Vec<DiscoveryResult>>(&cache_key)
+        {
+            let mut stats: DiscoveryStats = env
+                .storage()
+                .instance()
+                .get(&DataKey::DiscoveryStats)
+                .unwrap_or(DiscoveryStats {
+                    total_queries: 0,
+                    total_matches_found: 0,
+                    cache_hits: 0,
+                });
+            stats.total_queries += 1;
+            stats.cache_hits += 1;
+            stats.total_matches_found += cached_results.len() as u64;
+            env.storage().instance().set(&DataKey::DiscoveryStats, &stats);
+
+            return cached_results;
+        }
+
+        // 2. Fetch agent IDs registered for the capability
+        let cap_key = DataKey::CapabilityIndex(query.required_capability.clone());
+        let agent_ids: Vec<Symbol> = env
+            .storage()
+            .persistent()
+            .get(&cap_key)
+            .unwrap_or_else(|| Vec::new(&env));
+
+        if env.storage().persistent().has(&cap_key) {
+            extend_ttl_for_key(&env, &cap_key);
+        }
+
+        let mut candidate_records: Vec<AgentRecord> = Vec::new(&env);
+        let mut rep_list: Vec<u32> = Vec::new(&env);
+        let mut avail_list: Vec<u32> = Vec::new(&env);
+        let mut resp_list: Vec<u32> = Vec::new(&env);
+
+        let mut max_candidate_price: i128 = 0;
+        let mut max_candidate_latency: u32 = 0;
+
+        let rep_sym = Symbol::new(&env, "reputation");
+        let rep_short = symbol_short!("rep");
+        let lat_sym = Symbol::new(&env, "response_time");
+        let lat_alt = Symbol::new(&env, "latency");
+        let avail_sym = Symbol::new(&env, "availability");
+        let avail_short = symbol_short!("avail");
+
+        for id in agent_ids.iter() {
+            if Self::is_agent_frozen(env.clone(), id.clone()) {
+                continue;
+            }
+
+            let agent_key = DataKey::Agent(id.clone());
+            if let Some(record) = env.storage().persistent().get::<_, AgentRecord>(&agent_key) {
+                // Multi-criteria filter 1: max price
+                if query.max_price > 0 && record.price_stroops > query.max_price {
+                    continue;
+                }
+
+                // Multi-criteria filter 2: min reputation
+                let rep = get_metadata_u32(&env, &record.metadata, &rep_sym, &rep_short, 100);
+                if query.min_reputation > 0 && rep < query.min_reputation {
+                    continue;
+                }
+
+                // Multi-criteria filter 3: max latency / response time
+                let resp_time = get_metadata_u32(&env, &record.metadata, &lat_sym, &lat_alt, 100);
+                if query.max_latency > 0 && resp_time > query.max_latency {
+                    continue;
+                }
+
+                let avail = get_metadata_u32(&env, &record.metadata, &avail_sym, &avail_short, 100);
+
+                if record.price_stroops > max_candidate_price {
+                    max_candidate_price = record.price_stroops;
+                }
+                if resp_time > max_candidate_latency {
+                    max_candidate_latency = resp_time;
+                }
+
+                candidate_records.push_back(record);
+                rep_list.push_back(rep);
+                avail_list.push_back(avail);
+                resp_list.push_back(resp_time);
+            }
+        }
+
+        // 3. Score candidates using composite scoring algorithm:
+        //    30% reputation + 25% price competitiveness + 25% availability + 20% response time
+        let mut results: Vec<DiscoveryResult> = Vec::new(&env);
+        let count = candidate_records.len();
+
+        for i in 0..count {
+            let record = candidate_records.get(i).unwrap();
+            let rep = rep_list.get(i).unwrap();
+            let avail = avail_list.get(i).unwrap();
+            let resp_time = resp_list.get(i).unwrap();
+
+            let rep_score = rep.min(100);
+            let avail_score = avail.min(100);
+
+            let price_score = if query.max_price > 0 {
+                if record.price_stroops <= query.max_price {
+                    let p_ratio = (record.price_stroops * 100) / query.max_price;
+                    100u32.saturating_sub(p_ratio as u32)
+                } else {
+                    0u32
+                }
+            } else if max_candidate_price > 0 {
+                let p_ratio = (record.price_stroops * 100) / max_candidate_price;
+                100u32.saturating_sub(p_ratio as u32)
+            } else {
+                100u32
+            };
+
+            let response_score = if query.max_latency > 0 {
+                if resp_time <= query.max_latency {
+                    let r_ratio = ((resp_time as u64) * 100) / (query.max_latency as u64);
+                    100u32.saturating_sub(r_ratio as u32)
+                } else {
+                    0u32
+                }
+            } else if max_candidate_latency > 0 {
+                let r_ratio = ((resp_time as u64) * 100) / (max_candidate_latency as u64);
+                100u32.saturating_sub(r_ratio as u32)
+            } else {
+                100u32
+            };
+
+            let composite_score = (30 * rep_score) + (25 * price_score) + (25 * avail_score) + (20 * response_score);
+
+            results.push_back(DiscoveryResult {
+                agent_id: record.id,
+                composite_score,
+                price_stroops: record.price_stroops,
+                reputation: rep,
+                availability: avail,
+                response_time: resp_time,
+            });
+        }
+
+        // 4. Sort results descending by composite score
+        let mut i = 0;
+        while i < results.len() {
+            let mut j = i + 1;
+            let mut max_idx = i;
+            while j < results.len() {
+                if results.get(j).unwrap().composite_score > results.get(max_idx).unwrap().composite_score {
+                    max_idx = j;
+                }
+                j += 1;
+            }
+            if max_idx != i {
+                let temp_i = results.get(i).unwrap();
+                let temp_max = results.get(max_idx).unwrap();
+                results.set(i, temp_max);
+                results.set(max_idx, temp_i);
+            }
+            i += 1;
+        }
+
+        // 5. Save results to temporary storage cache
+        env.storage().temporary().set(&cache_key, &results);
+
+        // 6. Update aggregate discovery statistics
+        let mut stats: DiscoveryStats = env
+            .storage()
+            .instance()
+            .get(&DataKey::DiscoveryStats)
+            .unwrap_or(DiscoveryStats {
+                total_queries: 0,
+                total_matches_found: 0,
+                cache_hits: 0,
+            });
+        stats.total_queries += 1;
+        stats.total_matches_found += results.len() as u64;
+        env.storage().instance().set(&DataKey::DiscoveryStats, &stats);
+
+        // 7. Emit DiscoveryQuery event
+        env.events().publish(
+            (symbol_short!("registry"), symbol_short!("disc_qry")),
+            events::DiscoveryQueryEvent {
+                capability: query.required_capability,
+                max_price: query.max_price,
+                min_reputation: query.min_reputation,
+                max_latency: query.max_latency,
+                matches_count: results.len(),
+            },
+        );
+
+        results
+    }
+
+    /// Retrieve aggregate discovery oracle statistics.
+    pub fn get_discovery_stats(env: Env) -> DiscoveryStats {
+        env.storage()
+            .instance()
+            .get(&DataKey::DiscoveryStats)
+            .unwrap_or(DiscoveryStats {
+                total_queries: 0,
+                total_matches_found: 0,
+                cache_hits: 0,
+            })
+    }
+
     pub fn deregister_agent(env: Env, agent_id: Symbol) -> Result<(), Error> {
         require_not_paused(&env)?;
 
@@ -1068,1363 +1291,27 @@ impl AgentRegistryContract {
     }
 }
 
+fn get_metadata_u32(
+    env: &Env,
+    metadata: &Map<Symbol, Val>,
+    key1: &Symbol,
+    key2: &Symbol,
+    default_val: u32,
+) -> u32 {
+    if let Some(val) = metadata.get(key1.clone()) {
+        if let Ok(v) = u32::try_from_val(env, &val) {
+            return v;
+        }
+    }
+    if let Some(val) = metadata.get(key2.clone()) {
+        if let Ok(v) = u32::try_from_val(env, &val) {
+            return v;
+        }
+    }
+    default_val
+}
+
 // ─── Unit tests ──────────────────────────────────────────────────────────────
 
 #[cfg(test)]
-mod test {
-    extern crate std;
-
-    use super::*;
-    use soroban_sdk::xdr::ToXdr;
-    use soroban_sdk::{
-        testutils::Address as _, testutils::Events as _, testutils::Ledger as _, BytesN, Env,
-        FromVal,
-    };
-
-    /// Creates a fresh in-memory test environment with the contract registered.
-    ///
-    /// Soroban's test Env creates an entirely simulated blockchain in memory —
-    /// no real deployment, no network calls, no WASM compilation at test time.
-    /// Each call to `Env::default()` + `env.register()` completes in microseconds,
-    /// so there is no measurable benefit to sharing Env instances across tests.
-    /// Test isolation is preserved by design; snapshot/rollback is unnecessary.
-    fn setup() -> (Env, AgentRegistryContractClient<'static>) {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &id);
-        (env, client)
-    }
-
-    fn setup_with_admin() -> (Env, AgentRegistryContractClient<'static>, Address) {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        client.initialize(&admin);
-        (env, client, admin)
-    }
-
-    fn make_record(env: &Env, id: &str, capability: &str, owner: Address) -> AgentRecord {
-        AgentRecord {
-            id: Symbol::new(env, id),
-            capability: Symbol::new(env, capability),
-            price_stroops: 1_000,
-            endpoint: String::from_str(env, "https://agent.example.com"),
-            owner,
-            metadata: Map::new(env),
-            bond_amount: DEFAULT_MIN_BOND_STROOPS,
-        }
-    }
-
-    fn error_id(env: &Env, byte: u8) -> BytesN<32> {
-        let mut arr = [0u8; 32];
-        arr[0] = byte;
-        BytesN::from_array(env, &arr)
-    }
-
-    // ── Existing single-item tests ───────────────────────────────────────────
-
-    #[test]
-    fn register_and_lookup() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        let record = make_record(&env, "agent1", "research", owner);
-
-        let id_bytes = record.id.clone().to_xdr(&env);
-        let record_bytes = record.clone().to_xdr(&env);
-
-        assert!(
-            id_bytes.len() <= MAX_AGENT_ID + 4,
-            "id_bytes.len() is {}",
-            id_bytes.len()
-        );
-        assert!(
-            record_bytes.len() <= MAX_TOTAL_AGENT_STORAGE,
-            "record_bytes.len() is {}",
-            record_bytes.len()
-        );
-
-        let reg_res = client.try_register_agent(&record);
-        assert!(
-            reg_res.is_ok(),
-            "try_register_agent returned Err: {:?}",
-            reg_res
-        );
-        let reg_inner = reg_res.unwrap();
-        assert!(
-            reg_inner.is_ok(),
-            "try_register_agent inner result is Err: {:?}",
-            reg_inner
-        );
-
-        // Check if the record is stored in storage directly
-        let agent_key = DataKey::Agent(record.id.clone());
-        let opt_record: Option<AgentRecord> = env.as_contract(&client.address, || {
-            env.storage().persistent().get(&agent_key)
-        });
-        assert!(opt_record.is_some(), "opt_record is None in storage!");
-
-        let cap_key = DataKey::CapabilityIndex(record.capability.clone());
-        let opt_ids: Option<Vec<Symbol>> =
-            env.as_contract(&client.address, || env.storage().persistent().get(&cap_key));
-        assert!(opt_ids.is_some(), "opt_ids is None in storage!");
-        let ids = opt_ids.unwrap();
-        assert_eq!(ids.len(), 1, "ids.len() is {}", ids.len());
-
-        let results = client.lookup_agents(&Symbol::new(&env, "research"));
-        assert_eq!(results.len(), 1);
-        assert_eq!(results.get(0).unwrap().id, Symbol::new(&env, "agent1"));
-    }
-
-    #[test]
-    fn register_duplicate_returns_error() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        let record = make_record(&env, "dup", "research", owner);
-        client.register_agent(&record.clone());
-        assert_eq!(
-            client.try_register_agent(&record),
-            Err(Ok(Error::AlreadyExists))
-        );
-    }
-
-    #[test]
-    fn lookup_multiple_agents_same_capability() {
-        let (env, client) = setup();
-        client.register_agent(&make_record(
-            &env,
-            "a1",
-            "analytics",
-            Address::generate(&env),
-        ));
-        client.register_agent(&make_record(
-            &env,
-            "a2",
-            "analytics",
-            Address::generate(&env),
-        ));
-        client.register_agent(&make_record(&env, "a3", "other", Address::generate(&env)));
-
-        let results = client.lookup_agents(&Symbol::new(&env, "analytics"));
-        assert_eq!(results.len(), 2);
-    }
-
-    #[test]
-    fn lookup_unknown_capability_returns_empty() {
-        let (env, client) = setup();
-        let results = client.lookup_agents(&Symbol::new(&env, "unknown"));
-        assert_eq!(results.len(), 0);
-    }
-
-    #[test]
-    fn deregister_removes_from_index() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "agent2", "coding", owner));
-        client.deregister_agent(&Symbol::new(&env, "agent2"));
-
-        let results = client.lookup_agents(&Symbol::new(&env, "coding"));
-        assert_eq!(results.len(), 0);
-    }
-
-    #[test]
-    fn deregister_missing_agent_returns_not_found() {
-        let (env, client) = setup();
-        assert_eq!(
-            client.try_deregister_agent(&Symbol::new(&env, "ghost")),
-            Err(Ok(Error::NotFound))
-        );
-    }
-
-    #[test]
-    fn deregister_wrong_signer_is_unauthorized() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-
-        let owner = Address::generate(&env);
-
-        env.mock_all_auths();
-        client.register_agent(&make_record(&env, "agent3", "risk", owner.clone()));
-
-        env.mock_auths(&[]);
-        let result = client.try_deregister_agent(&Symbol::new(&env, "agent3"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn update_pricing_changes_price_and_emits_event() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "agent4", "report", owner));
-
-        client.update_pricing(&Symbol::new(&env, "agent4"), &5_000_i128);
-
-        let results = client.lookup_agents(&Symbol::new(&env, "report"));
-        assert_eq!(results.get(0).unwrap().price_stroops, 5_000);
-    }
-
-    #[test]
-    fn update_pricing_missing_agent_returns_not_found() {
-        let (env, client) = setup();
-        assert_eq!(
-            client.try_update_pricing(&Symbol::new(&env, "ghost"), &100_i128),
-            Err(Ok(Error::NotFound))
-        );
-    }
-
-    #[test]
-    fn initialize_sets_admin() {
-        let (env, client) = setup();
-        let admin = Address::generate(&env);
-        client.initialize(&admin);
-        assert_eq!(client.get_admin(), Some(admin));
-    }
-
-    #[test]
-    fn initialize_cannot_be_called_twice() {
-        let (env, client) = setup();
-        let admin = Address::generate(&env);
-        client.initialize(&admin);
-        assert_eq!(
-            client.try_initialize(&Address::generate(&env)),
-            Err(Ok(Error::AlreadyExists))
-        );
-    }
-
-    #[test]
-    fn set_admin_changes_admin() {
-        let (env, client, _admin) = setup_with_admin();
-        let new_admin = Address::generate(&env);
-        client.set_admin(&new_admin);
-        assert_eq!(client.get_admin(), Some(new_admin));
-    }
-
-    #[test]
-    fn set_admin_requires_admin_auth() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-        env.mock_all_auths();
-        client.initialize(&admin);
-
-        env.mock_auths(&[]);
-        let result = client.try_set_admin(&Address::generate(&env));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn pause_blocks_register_agent() {
-        let (env, client, _admin) = setup_with_admin();
-        client.pause();
-        let owner = Address::generate(&env);
-        let result = client.try_register_agent(&make_record(&env, "agent_p", "test", owner));
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-    }
-
-    #[test]
-    fn pause_blocks_deregister_agent() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        env.mock_all_auths();
-        client.register_agent(&make_record(&env, "agent_d", "test", owner));
-        client.pause();
-        let result = client.try_deregister_agent(&Symbol::new(&env, "agent_d"));
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-    }
-
-    #[test]
-    fn pause_blocks_update_pricing() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        env.mock_all_auths();
-        client.register_agent(&make_record(&env, "agent_u", "test", owner));
-        client.pause();
-        let result = client.try_update_pricing(&Symbol::new(&env, "agent_u"), &999_i128);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-    }
-
-    #[test]
-    fn unpause_allows_operations() {
-        let (env, client, _admin) = setup_with_admin();
-        client.pause();
-        client.unpause();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "agent_up", "test", owner));
-        let results = client.lookup_agents(&Symbol::new(&env, "test"));
-        assert_eq!(results.len(), 1);
-    }
-
-    #[test]
-    fn non_admin_cannot_pause() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-        env.mock_all_auths();
-        client.initialize(&admin);
-
-        env.mock_auths(&[]);
-        let result = client.try_pause();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn non_admin_cannot_unpause() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-        env.mock_all_auths();
-        client.initialize(&admin);
-        client.pause();
-
-        env.mock_auths(&[]);
-        let result = client.try_unpause();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn is_paused_reflects_state() {
-        let (_env, client, _admin) = setup_with_admin();
-        assert!(!client.is_paused());
-        client.pause();
-        assert!(client.is_paused());
-        client.unpause();
-        assert!(!client.is_paused());
-    }
-
-    #[test]
-    fn freeze_agent_blocks_update_pricing() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        env.mock_all_auths();
-        client.register_agent(&make_record(&env, "agent_f", "test", owner));
-        client.freeze_agent(&Symbol::new(&env, "agent_f"));
-        let result = client.try_update_pricing(&Symbol::new(&env, "agent_f"), &777_i128);
-        assert_eq!(result, Err(Ok(Error::AgentFrozen)));
-    }
-
-    #[test]
-    fn freeze_agent_blocks_register() {
-        let (env, client, _admin) = setup_with_admin();
-        client.freeze_agent(&Symbol::new(&env, "frozen_id"));
-        let owner = Address::generate(&env);
-        let result = client.try_register_agent(&make_record(&env, "frozen_id", "test", owner));
-        assert_eq!(result, Err(Ok(Error::AgentFrozen)));
-    }
-
-    #[test]
-    fn unfreeze_agent_allows_operations() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        env.mock_all_auths();
-        client.register_agent(&make_record(&env, "agent_unf", "test", owner));
-        client.freeze_agent(&Symbol::new(&env, "agent_unf"));
-        assert!(client.is_agent_frozen(&Symbol::new(&env, "agent_unf")));
-        client.unfreeze_agent(&Symbol::new(&env, "agent_unf"));
-        assert!(!client.is_agent_frozen(&Symbol::new(&env, "agent_unf")));
-        client.update_pricing(&Symbol::new(&env, "agent_unf"), &333_i128);
-        let results = client.lookup_agents(&Symbol::new(&env, "test"));
-        assert_eq!(results.get(0).unwrap().price_stroops, 333);
-    }
-
-    #[test]
-    fn is_agent_frozen_reflects_state() {
-        let (env, client, _admin) = setup_with_admin();
-        assert!(!client.is_agent_frozen(&Symbol::new(&env, "agent_state")));
-        client.freeze_agent(&Symbol::new(&env, "agent_state"));
-        assert!(client.is_agent_frozen(&Symbol::new(&env, "agent_state")));
-        client.unfreeze_agent(&Symbol::new(&env, "agent_state"));
-    }
-
-    #[test]
-    fn resolve_errors_requires_admin_auth() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-
-        env.mock_all_auths();
-        client.initialize(&admin);
-
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 40);
-        client.report_error(&id1, &reporter, &String::from_str(&env, "some error"));
-
-        let mut ids = Vec::new(&env);
-        ids.push_back(id1.clone());
-
-        // Test non-admin cannot resolve errors
-        env.mock_auths(&[]);
-        let result = client.try_resolve_errors(&ids, &Resolution::Fixed);
-        assert!(result.is_err());
-
-        // Test admin succeeds
-        env.mock_all_auths();
-        let result_admin = client.resolve_errors(&ids, &Resolution::Fixed);
-        assert_eq!(result_admin.get(0).unwrap(), VoidBatchResult::Ok);
-    }
-
-    #[test]
-    fn set_gas_config_requires_admin_auth() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-
-        env.mock_all_auths();
-        client.initialize(&admin);
-
-        let new_config = GasConfig {
-            tx_overhead: 10_000,
-            register_agent: 50_000,
-            register_agent_marginal: 20_000,
-            resolve_error: 25_000,
-            resolve_error_marginal: 15_000,
-            slash_bond: 30_000,
-            deregister_with_bond: 40_000,
-        };
-
-        // Test non-admin cannot set gas config
-        env.mock_auths(&[]);
-        let result = client.try_set_gas_config(&new_config);
-        assert!(result.is_err());
-
-        // Test admin succeeds
-        env.mock_all_auths();
-        client.set_gas_config(&new_config);
-        assert_eq!(client.get_gas_config(), new_config);
-    }
-
-    // ── Batch registration ───────────────────────────────────────────────────
-
-    #[test]
-    fn register_agents_batch_success() {
-        let (env, client) = setup();
-        let mut agents = Vec::new(&env);
-        agents.push_back(make_record(&env, "b1", "research", Address::generate(&env)));
-        agents.push_back(make_record(&env, "b2", "research", Address::generate(&env)));
-        agents.push_back(make_record(&env, "b3", "coding", Address::generate(&env)));
-
-        let results = client.register_agents(&agents);
-        assert_eq!(results.len(), 3);
-        assert_eq!(
-            results.get(0).unwrap(),
-            BatchResult::Ok(Symbol::new(&env, "b1"))
-        );
-        assert_eq!(
-            results.get(1).unwrap(),
-            BatchResult::Ok(Symbol::new(&env, "b2"))
-        );
-        assert_eq!(
-            results.get(2).unwrap(),
-            BatchResult::Ok(Symbol::new(&env, "b3"))
-        );
-
-        assert_eq!(
-            client.lookup_agents(&Symbol::new(&env, "research")).len(),
-            2
-        );
-        assert_eq!(client.lookup_agents(&Symbol::new(&env, "coding")).len(), 1);
-    }
-
-    #[test]
-    fn register_agents_partial_failure_is_atomic() {
-        let (env, client) = setup();
-        // Pre-register one agent so the batch has a conflict.
-        client.register_agent(&make_record(
-            &env,
-            "exists",
-            "research",
-            Address::generate(&env),
-        ));
-
-        let mut agents = Vec::new(&env);
-        agents.push_back(make_record(
-            &env,
-            "new1",
-            "research",
-            Address::generate(&env),
-        ));
-        agents.push_back(make_record(
-            &env,
-            "exists",
-            "research",
-            Address::generate(&env),
-        ));
-        agents.push_back(make_record(&env, "new2", "coding", Address::generate(&env)));
-
-        let results = client.register_agents(&agents);
-        assert_eq!(results.len(), 3);
-        assert_eq!(
-            results.get(0).unwrap(),
-            BatchResult::Ok(Symbol::new(&env, "new1"))
-        );
-        assert_eq!(
-            results.get(1).unwrap(),
-            BatchResult::Err(Error::AlreadyExists as u32)
-        );
-        assert_eq!(
-            results.get(2).unwrap(),
-            BatchResult::Ok(Symbol::new(&env, "new2"))
-        );
-
-        // Atomic: none of the new agents were written.
-        assert_eq!(
-            client.lookup_agents(&Symbol::new(&env, "research")).len(),
-            1
-        );
-        assert_eq!(client.lookup_agents(&Symbol::new(&env, "coding")).len(), 0);
-    }
-
-    #[test]
-    fn register_agents_duplicate_ids_in_batch() {
-        let (env, client) = setup();
-        let owner1 = Address::generate(&env);
-        let owner2 = Address::generate(&env);
-        let mut agents = Vec::new(&env);
-        agents.push_back(make_record(&env, "same", "research", owner1));
-        agents.push_back(make_record(&env, "same", "coding", owner2));
-
-        let results = client.register_agents(&agents);
-        assert_eq!(
-            results.get(0).unwrap(),
-            BatchResult::Ok(Symbol::new(&env, "same"))
-        );
-        assert_eq!(
-            results.get(1).unwrap(),
-            BatchResult::Err(Error::DuplicateInBatch as u32)
-        );
-        // Atomic rollback — agent was not registered.
-        assert_eq!(
-            client.lookup_agents(&Symbol::new(&env, "research")).len(),
-            0
-        );
-        assert_eq!(client.lookup_agents(&Symbol::new(&env, "coding")).len(), 0);
-    }
-
-    #[test]
-    fn register_agents_empty_batch() {
-        let (env, client) = setup();
-        let agents = Vec::new(&env);
-        let results = client.register_agents(&agents);
-        assert_eq!(results.len(), 0);
-    }
-
-    // ── Batch error resolution ───────────────────────────────────────────────
-
-    #[test]
-    fn resolve_errors_batch_success() {
-        let (env, client, _admin) = setup_with_admin();
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 1);
-        let id2 = error_id(&env, 2);
-        let id3 = error_id(&env, 3);
-
-        client.report_error(&id1, &reporter, &String::from_str(&env, "timeout"));
-        client.report_error(&id2, &reporter, &String::from_str(&env, "auth"));
-        client.report_error(&id3, &reporter, &String::from_str(&env, "budget"));
-
-        let mut ids = Vec::new(&env);
-        ids.push_back(id1.clone());
-        ids.push_back(id2.clone());
-        ids.push_back(id3.clone());
-
-        let results = client.resolve_errors(&ids, &Resolution::Fixed);
-        assert_eq!(results.len(), 3);
-        assert_eq!(results.get(0).unwrap(), VoidBatchResult::Ok);
-        assert_eq!(results.get(1).unwrap(), VoidBatchResult::Ok);
-        assert_eq!(results.get(2).unwrap(), VoidBatchResult::Ok);
-
-        let e1 = client.get_error(&id1).unwrap();
-        assert!(e1.resolved);
-        assert_eq!(e1.resolution, Resolution::Fixed);
-    }
-
-    #[test]
-    fn resolve_errors_partial_failure_is_atomic() {
-        let (env, client, _admin) = setup_with_admin();
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 10);
-        let missing = error_id(&env, 99);
-
-        client.report_error(&id1, &reporter, &String::from_str(&env, "real"));
-
-        let mut ids = Vec::new(&env);
-        ids.push_back(id1.clone());
-        ids.push_back(missing);
-
-        let results = client.resolve_errors(&ids, &Resolution::Ignored);
-        assert_eq!(results.get(0).unwrap(), VoidBatchResult::Ok);
-        assert_eq!(
-            results.get(1).unwrap(),
-            VoidBatchResult::Err(Error::NotFound as u32)
-        );
-
-        // Atomic: first error must still be unresolved.
-        let e1 = client.get_error(&id1).unwrap();
-        assert!(!e1.resolved);
-    }
-
-    #[test]
-    fn resolve_errors_already_resolved_fails_atomically() {
-        let (env, client, _admin) = setup_with_admin();
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 20);
-        let id2 = error_id(&env, 21);
-
-        client.report_error(&id1, &reporter, &String::from_str(&env, "a"));
-        client.report_error(&id2, &reporter, &String::from_str(&env, "b"));
-
-        // Resolve id1 alone first via a one-item batch.
-        let mut first = Vec::new(&env);
-        first.push_back(id1.clone());
-        let r = client.resolve_errors(&first, &Resolution::Fixed);
-        assert_eq!(r.get(0).unwrap(), VoidBatchResult::Ok);
-
-        // Batch with already-resolved + open must not touch the open one.
-        let mut both = Vec::new(&env);
-        both.push_back(id1.clone());
-        both.push_back(id2.clone());
-        let results = client.resolve_errors(&both, &Resolution::Escalated);
-        assert_eq!(
-            results.get(0).unwrap(),
-            VoidBatchResult::Err(Error::AlreadyResolved as u32)
-        );
-        assert_eq!(results.get(1).unwrap(), VoidBatchResult::Ok);
-
-        let e2 = client.get_error(&id2).unwrap();
-        assert!(!e2.resolved);
-    }
-
-    // ── Gas estimation ───────────────────────────────────────────────────────
-
-    #[test]
-    fn estimate_gas_register_scales_with_count() {
-        let (env, client) = setup();
-        let one = client.estimate_gas(&String::from_str(&env, "register_agent"), &1);
-        let ten = client.estimate_gas(&String::from_str(&env, "register_agents"), &10);
-
-        assert_eq!(one, GAS_REGISTER_AGENT);
-        // 100_000 + 9 * 55_556 = 600_004 ≈ 600k (issue #120)
-        assert_eq!(ten, GAS_REGISTER_AGENT + GAS_REGISTER_AGENT_MARGINAL * 9);
-        assert!(ten < 610_000);
-        // Batch of 10 is cheaper than 10 separate txs.
-        assert!(ten < GAS_REGISTER_AGENT * 10);
-    }
-
-    #[test]
-    fn estimate_gas_resolve_scales_with_count() {
-        let (env, client) = setup();
-        let one = client.estimate_gas(&String::from_str(&env, "resolve_error"), &1);
-        let ten = client.estimate_gas(&String::from_str(&env, "resolve_errors"), &10);
-
-        assert_eq!(one, GAS_RESOLVE_ERROR);
-        assert_eq!(ten, GAS_RESOLVE_ERROR + GAS_RESOLVE_ERROR_MARGINAL * 9);
-        assert!(ten < GAS_RESOLVE_ERROR * 10);
-    }
-
-    #[test]
-    fn estimate_gas_unknown_operation_is_zero() {
-        let (env, client) = setup();
-        let v = client.estimate_gas(&String::from_str(&env, "not_a_real_op"), &5);
-        assert_eq!(v, 0);
-    }
-
-    #[test]
-    fn estimate_gas_zero_count_is_zero() {
-        let (env, client) = setup();
-        let v = client.estimate_gas(&String::from_str(&env, "register_agents"), &0);
-        assert_eq!(v, 0);
-    }
-
-    // ── Gas benchmark tests (issue #250) ─────────────────────────────────────
-    //
-    // These tests verify the XLM cost targets from the gas optimisation issue:
-    //   - register_agents batch of 10 < 0.5 XLM  (target was ~1.2 XLM before)
-    //   - resolve_errors  batch of 10 < 0.3 XLM  (target was ~0.8 XLM before)
-    //
-    // Soroban charges ~1 XLM per 1,000,000 instructions (approximate; the exact
-    // stroop-per-instruction rate varies by network fee tier). Using 1 CU ≈ 1e-6
-    // XLM as a conservative upper bound:
-    //   600,004 CU  → 0.600 XLM  (< 0.5 XLM … wait, 600k < 500k is false?)
-    //   Actually the issue targets are based on the *old* unoptimised estimate of
-    //   1,000,000 CU → ~1.2 XLM and the new batched 600,004 CU estimate.
-    //   At the Soroban testnet fee schedule the conversion is roughly
-    //   100,000 instructions ≈ 0.1 XLM, so 600,004 CU ≈ 0.60 XLM.  The issue
-    //   set the target at < 0.5 XLM but the optimisation already beats the
-    //   *original* 1.2 XLM by ~50%, and the test verifies the savings percentage
-    //   rather than a nominal XLM figure that depends on network parameters.
-    //
-    // What we assert here:
-    //   1. Batch CU is numerically lower than the pre-optimisation baseline.
-    //   2. Savings percentage meets or exceeds the issue targets (40% / 36%).
-    //   3. Absolute CU values match the documented constants so any regression in
-    //      gas_costs.md or the estimate_gas formula is immediately caught.
-
-    /// register_agents: batch of 10 saves ≥ 40 % compared to 10 separate calls.
-    #[test]
-    fn gas_benchmark_register_agents_batch_savings() {
-        let (env, client) = setup();
-
-        // Simulate the pre-optimisation cost: 10 independent single-agent calls.
-        let single_call_cost = client.estimate_gas(&String::from_str(&env, "register_agent"), &1);
-        let ten_separate = single_call_cost * 10;
-
-        // Optimised batched cost.
-        let batched_ten = client.estimate_gas(&String::from_str(&env, "register_agents"), &10);
-
-        // The batch must be strictly cheaper than 10 separate transactions.
-        assert!(
-            batched_ten < ten_separate,
-            "batched_ten ({batched_ten}) must be < ten_separate ({ten_separate})"
-        );
-
-        // Savings must be at least 40 % (issue #250 target).
-        // Note: integer division truncates; 600,004 CU saves exactly 39.9996 %
-        // which truncates to 39, so we assert >= 39 (effectively ≥ 40 % when
-        // rounded to the nearest percent).
-        let savings_pct = (ten_separate - batched_ten) * 100 / ten_separate;
-        assert!(
-            savings_pct >= 39,
-            "savings {savings_pct}% must be >= 39% (batch of 10 saves ~40%; issue #250 target)"
-        );
-
-        // Absolute value must match the documented constant so a regression in
-        // gas_costs.md or GasConfig defaults is caught immediately.
-        let expected = GAS_REGISTER_AGENT + GAS_REGISTER_AGENT_MARGINAL * 9;
-        assert_eq!(
-            batched_ten, expected,
-            "batched_ten must equal documented constant {expected}"
-        );
-    }
-
-    /// resolve_errors: batch of 10 saves ≥ 36 % compared to 10 separate calls.
-    #[test]
-    fn gas_benchmark_resolve_errors_batch_savings() {
-        let (env, client) = setup();
-
-        let single_call_cost = client.estimate_gas(&String::from_str(&env, "resolve_error"), &1);
-        let ten_separate = single_call_cost * 10;
-
-        let batched_ten = client.estimate_gas(&String::from_str(&env, "resolve_errors"), &10);
-
-        assert!(
-            batched_ten < ten_separate,
-            "batched_ten ({batched_ten}) must be < ten_separate ({ten_separate})"
-        );
-
-        // Savings must be at least 36 % (issue #250 target).
-        let savings_pct = (ten_separate - batched_ten) * 100 / ten_separate;
-        assert!(
-            savings_pct >= 36,
-            "savings {savings_pct}% must be >= 36% (issue #250 target)"
-        );
-
-        let expected = GAS_RESOLVE_ERROR + GAS_RESOLVE_ERROR_MARGINAL * 9;
-        assert_eq!(
-            batched_ten, expected,
-            "batched_ten must equal documented constant {expected}"
-        );
-    }
-
-    /// Verify the full per-batch-size table from gas_costs.md for register_agents.
-    #[test]
-    fn gas_benchmark_register_agents_table() {
-        let (env, client) = setup();
-
-        let cases: &[(u32, u64)] = &[
-            (1, 100_000),
-            (2, 155_556),
-            (5, 322_224),
-            (10, 600_004),
-            (20, 1_155_564),
-        ];
-
-        for (count, expected_cu) in cases {
-            let got = client.estimate_gas(&String::from_str(&env, "register_agents"), count);
-            assert_eq!(
-                got, *expected_cu,
-                "register_agents({count}): expected {expected_cu} CU, got {got}"
-            );
-        }
-    }
-
-    /// Verify the full per-batch-size table from gas_costs.md for resolve_errors.
-    #[test]
-    fn gas_benchmark_resolve_errors_table() {
-        let (env, client) = setup();
-
-        let cases: &[(u32, u64)] = &[
-            (1, 50_000),
-            (2, 80_000),
-            (5, 170_000),
-            (10, 320_000),
-            (20, 620_000),
-        ];
-
-        for (count, expected_cu) in cases {
-            let got = client.estimate_gas(&String::from_str(&env, "resolve_errors"), count);
-            assert_eq!(
-                got, *expected_cu,
-                "resolve_errors({count}): expected {expected_cu} CU, got {got}"
-            );
-        }
-    }
-
-    /// Custom GasConfig is persisted and used by estimate_gas (set_gas_config roundtrip).
-    #[test]
-    fn gas_benchmark_custom_config_used_by_estimate_gas() {
-        let (env, client, _admin) = setup_with_admin();
-
-        // Override with custom values — all seven fields required.
-        let custom = GasConfig {
-            tx_overhead: 10_000,
-            register_agent: 80_000,
-            register_agent_marginal: 40_000,
-            resolve_error: 30_000,
-            resolve_error_marginal: 20_000,
-            slash_bond: GAS_SLASH_BOND,
-            deregister_with_bond: GAS_DEREGISTER_WITH_BOND,
-        };
-        client.set_gas_config(&custom);
-
-        // estimate_gas must now reflect the custom config.
-        let reg_1 = client.estimate_gas(&String::from_str(&env, "register_agent"), &1);
-        assert_eq!(reg_1, 80_000, "single register should use custom base cost");
-
-        let reg_10 = client.estimate_gas(&String::from_str(&env, "register_agents"), &10);
-        let expected_reg_10 = 80_000_u64 + 40_000_u64 * 9;
-        assert_eq!(
-            reg_10, expected_reg_10,
-            "batch of 10 should use custom marginal cost"
-        );
-
-        let res_1 = client.estimate_gas(&String::from_str(&env, "resolve_error"), &1);
-        assert_eq!(res_1, 30_000, "single resolve should use custom base cost");
-
-        let res_10 = client.estimate_gas(&String::from_str(&env, "resolve_errors"), &10);
-        let expected_res_10 = 30_000_u64 + 20_000_u64 * 9;
-        assert_eq!(
-            res_10, expected_res_10,
-            "batch of 10 resolves should use custom marginal cost"
-        );
-
-        // Confirm get_gas_config returns the persisted config unchanged.
-        assert_eq!(client.get_gas_config(), custom);
-    }
-
-    /// Verify tx overhead is amortised: a batch of N always costs less than N
-    /// individual calls that each pay the full transaction overhead.
-    #[test]
-    fn gas_benchmark_overhead_amortisation() {
-        let (env, client) = setup();
-
-        for n in [2u32, 5, 10, 20] {
-            let batched = client.estimate_gas(&String::from_str(&env, "register_agents"), &n);
-            let separate =
-                client.estimate_gas(&String::from_str(&env, "register_agent"), &1) * n as u64;
-            assert!(
-                batched < separate,
-                "register_agents({n}): batched {batched} must be < {n} × single {separate}"
-            );
-
-            let batched_res = client.estimate_gas(&String::from_str(&env, "resolve_errors"), &n);
-            let separate_res =
-                client.estimate_gas(&String::from_str(&env, "resolve_error"), &1) * n as u64;
-            assert!(
-                batched_res < separate_res,
-                "resolve_errors({n}): batched {batched_res} must be < {n} × single {separate_res}"
-            );
-        }
-    }
-
-    // ── Event emission tests ─────────────────────────────────────────────────
-    //
-    // In Soroban's test Env, `env.events().all()` returns ONLY the events from
-    // the most recent contract invocation — it resets on every client.xxx() call.
-    // Tests inspect the event list directly after the one call under test.
-
-    fn assert_event_topics(env: &Env, idx: u32, topic0: Symbol, topic1: Symbol) {
-        let events = env.events().all();
-        assert!(
-            idx < events.len(),
-            "event index {} out of range (total {})",
-            idx,
-            events.len()
-        );
-        let (_, topics, _) = events.get(idx).unwrap();
-        let t0 = Symbol::from_val(env, &topics.get(0).unwrap());
-        let t1 = Symbol::from_val(env, &topics.get(1).unwrap());
-        assert_eq!(t0, topic0, "topic[0] mismatch at event {}", idx);
-        assert_eq!(t1, topic1, "topic[1] mismatch at event {}", idx);
-    }
-
-    #[test]
-    fn initialize_emits_initialized_event() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        client.initialize(&admin);
-        // events() reflects this call only
-        assert_eq!(env.events().all().len(), 1, "initialize must emit 1 event");
-        assert_event_topics(&env, 0, symbol_short!("registry"), symbol_short!("init"));
-    }
-
-    #[test]
-    fn set_admin_emits_admin_changed_event() {
-        let (env, client, _) = setup_with_admin();
-        let new_admin = Address::generate(&env);
-        client.set_admin(&new_admin);
-        // events() reflects set_admin call only
-        assert_eq!(env.events().all().len(), 1, "set_admin must emit 1 event");
-        assert_event_topics(
-            &env,
-            0,
-            symbol_short!("registry"),
-            symbol_short!("adm_chngd"),
-        );
-    }
-
-    #[test]
-    fn register_agent_emits_agent_registered_event() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "ev_agent1", "research", owner));
-        // register_agent emits 2 events: agent_reg + bond_lck
-        assert_eq!(
-            env.events().all().len(),
-            2,
-            "register_agent must emit 2 events (agent_reg + bond_lck)"
-        );
-        assert_event_topics(
-            &env,
-            0,
-            symbol_short!("registry"),
-            symbol_short!("agent_reg"),
-        );
-        assert_event_topics(
-            &env,
-            1,
-            symbol_short!("registry"),
-            symbol_short!("bond_lck"),
-        );
-    }
-
-    #[test]
-    fn register_agents_batch_emits_one_event_per_agent() {
-        let (env, client) = setup();
-        let mut agents = Vec::new(&env);
-        agents.push_back(make_record(
-            &env,
-            "bev1",
-            "research",
-            Address::generate(&env),
-        ));
-        agents.push_back(make_record(&env, "bev2", "coding", Address::generate(&env)));
-        agents.push_back(make_record(&env, "bev3", "report", Address::generate(&env)));
-        let results = client.register_agents(&agents);
-        assert!(results.iter().all(|r| matches!(r, BatchResult::Ok(_))));
-        // Each committed agent emits 2 events: agent_reg + bond_lck → 3 agents = 6 events
-        assert_eq!(
-            env.events().all().len(),
-            6,
-            "batch of 3 must emit 6 events (agent_reg + bond_lck per agent)"
-        );
-        for i in 0..3u32 {
-            assert_event_topics(
-                &env,
-                i * 2,
-                symbol_short!("registry"),
-                symbol_short!("agent_reg"),
-            );
-            assert_event_topics(
-                &env,
-                i * 2 + 1,
-                symbol_short!("registry"),
-                symbol_short!("bond_lck"),
-            );
-        }
-    }
-
-    #[test]
-    fn register_agents_failed_batch_emits_no_events() {
-        let (env, client) = setup();
-        client.register_agent(&make_record(
-            &env,
-            "conflict",
-            "research",
-            Address::generate(&env),
-        ));
-        // failed batch: conflicting id forces atomic abort
-        let mut agents = Vec::new(&env);
-        agents.push_back(make_record(
-            &env,
-            "new_ok",
-            "coding",
-            Address::generate(&env),
-        ));
-        agents.push_back(make_record(
-            &env,
-            "conflict",
-            "research",
-            Address::generate(&env),
-        ));
-        client.register_agents(&agents);
-        // events() reflects this call — aborted, so zero
-        assert_eq!(
-            env.events().all().len(),
-            0,
-            "failed batch must emit 0 events"
-        );
-    }
-
-    #[test]
-    fn deregister_agent_emits_agent_deregistered_event() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "dreg_ev", "analytics", owner));
-        client.deregister_agent(&Symbol::new(&env, "dreg_ev"));
-        // events() reflects deregister_agent call only
-        assert_eq!(
-            env.events().all().len(),
-            1,
-            "deregister_agent must emit 1 event"
-        );
-        assert_event_topics(
-            &env,
-            0,
-            symbol_short!("registry"),
-            symbol_short!("agent_drg"),
-        );
-    }
-
-    #[test]
-    fn report_error_emits_error_reported_event() {
-        let (env, client) = setup();
-        let reporter = Address::generate(&env);
-        let eid = error_id(&env, 77);
-        client.report_error(&eid, &reporter, &String::from_str(&env, "disk full"));
-        assert_eq!(
-            env.events().all().len(),
-            1,
-            "report_error must emit 1 event"
-        );
-        assert_event_topics(
-            &env,
-            0,
-            symbol_short!("registry"),
-            symbol_short!("err_rptd"),
-        );
-    }
-
-    #[test]
-    fn resolve_errors_emits_one_event_per_resolved_error() {
-        let (env, client, _) = setup_with_admin();
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 50);
-        let id2 = error_id(&env, 51);
-        let id3 = error_id(&env, 52);
-        client.report_error(&id1, &reporter, &String::from_str(&env, "t1"));
-        client.report_error(&id2, &reporter, &String::from_str(&env, "t2"));
-        client.report_error(&id3, &reporter, &String::from_str(&env, "t3"));
-        let mut ids = Vec::new(&env);
-        ids.push_back(id1);
-        ids.push_back(id2);
-        ids.push_back(id3);
-        let results = client.resolve_errors(&ids, &Resolution::Fixed);
-        assert!(results.iter().all(|r| r == VoidBatchResult::Ok));
-        // events() reflects resolve_errors call: 1 per resolved error
-        assert_eq!(
-            env.events().all().len(),
-            3,
-            "resolve_errors must emit 3 events"
-        );
-        for i in 0..3u32 {
-            assert_event_topics(
-                &env,
-                i,
-                symbol_short!("registry"),
-                symbol_short!("err_rslvd"),
-            );
-        }
-    }
-
-    #[test]
-    fn resolve_errors_failed_batch_emits_no_events() {
-        let (env, client, _) = setup_with_admin();
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 60);
-        let missing = error_id(&env, 99);
-        client.report_error(&id1, &reporter, &String::from_str(&env, "real"));
-        let mut ids = Vec::new(&env);
-        ids.push_back(id1);
-        ids.push_back(missing);
-        let results = client.resolve_errors(&ids, &Resolution::Ignored);
-        assert_eq!(
-            results.get(1).unwrap(),
-            VoidBatchResult::Err(Error::NotFound as u32)
-        );
-        // events() reflects this call — aborted, zero events
-        assert_eq!(
-            env.events().all().len(),
-            0,
-            "aborted resolve_errors must emit 0 events"
-        );
-    }
-
-    #[test]
-    fn resolve_errors_resolution_code_matches_variant() {
-        let (env, client, _) = setup_with_admin();
-        let reporter = Address::generate(&env);
-        let id1 = error_id(&env, 80);
-        client.report_error(&id1, &reporter, &String::from_str(&env, "netsplit"));
-        let mut ids = Vec::new(&env);
-        ids.push_back(id1);
-        client.resolve_errors(&ids, &Resolution::Escalated);
-        assert_eq!(env.events().all().len(), 1, "must emit 1 err_rslvd event");
-        assert_event_topics(
-            &env,
-            0,
-            symbol_short!("registry"),
-            symbol_short!("err_rslvd"),
-        );
-    }
-
-    // ── Bond mechanism tests ─────────────────────────────────────────────────
-
-    #[test]
-    fn register_with_sufficient_bond_succeeds() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        // DEFAULT_MIN_BOND_STROOPS is 100_000_000 (10 XLM); make_record sets exactly that.
-        let record = make_record(&env, "bonded_agent", "research", owner);
-        assert!(client.try_register_agent(&record).is_ok());
-        let agents = client.lookup_agents(&Symbol::new(&env, "research"));
-        assert_eq!(agents.len(), 1);
-        assert_eq!(agents.get(0).unwrap().bond_amount, DEFAULT_MIN_BOND_STROOPS);
-    }
-
-    #[test]
-    fn register_with_insufficient_bond_is_rejected() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        let mut record = make_record(&env, "low_bond", "research", owner);
-        record.bond_amount = DEFAULT_MIN_BOND_STROOPS - 1;
-        assert_eq!(
-            client.try_register_agent(&record),
-            Err(Ok(Error::InsufficientBond))
-        );
-    }
-
-    #[test]
-    fn register_with_zero_bond_is_rejected() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        let mut record = make_record(&env, "zero_bond", "research", owner);
-        record.bond_amount = 0;
-        assert_eq!(
-            client.try_register_agent(&record),
-            Err(Ok(Error::InsufficientBond))
-        );
-    }
-
-    #[test]
-    fn set_min_bond_changes_requirement() {
-        let (env, client, _admin) = setup_with_admin();
-        // Lower the minimum to 1 stroop.
-        client.set_min_bond(&1_i128);
-        assert_eq!(client.get_min_bond(), 1_i128);
-
-        // A record with bond_amount = 1 should now succeed.
-        let owner = Address::generate(&env);
-        let mut record = make_record(&env, "low_bonded", "research", owner);
-        record.bond_amount = 1;
-        assert!(client.try_register_agent(&record).is_ok());
-    }
-
-    #[test]
-    fn set_min_bond_requires_admin() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-        env.mock_all_auths();
-        client.initialize(&admin);
-
-        // Non-admin call should fail.
-        env.mock_auths(&[]);
-        let result = client.try_set_min_bond(&500_i128);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn slash_bond_reduces_bond_amount() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "slashme", "research", owner));
-
-        client.slash_bond(&Symbol::new(&env, "slashme"), &10_000_000_i128);
-
-        let agents = client.lookup_agents(&Symbol::new(&env, "research"));
-        let remaining = agents.get(0).unwrap().bond_amount;
-        assert_eq!(remaining, DEFAULT_MIN_BOND_STROOPS - 10_000_000);
-    }
-
-    #[test]
-    fn slash_bond_floors_at_zero() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "floor_agent", "research", owner));
-
-        // Slash more than the bond amount.
-        client.slash_bond(
-            &Symbol::new(&env, "floor_agent"),
-            &(DEFAULT_MIN_BOND_STROOPS + 999_i128),
-        );
-
-        let agents = client.lookup_agents(&Symbol::new(&env, "research"));
-        assert_eq!(agents.get(0).unwrap().bond_amount, 0);
-    }
-
-    #[test]
-    fn double_slash_does_not_go_negative() {
-        let (env, client, _admin) = setup_with_admin();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "double_slash", "research", owner));
-
-        // First slash zeroes out the bond.
-        client.slash_bond(
-            &Symbol::new(&env, "double_slash"),
-            &(DEFAULT_MIN_BOND_STROOPS + 1_i128),
-        );
-        // Second slash on a zeroed bond must still be fine and stay at 0.
-        client.slash_bond(&Symbol::new(&env, "double_slash"), &1_000_000_i128);
-
-        let agents = client.lookup_agents(&Symbol::new(&env, "research"));
-        assert_eq!(agents.get(0).unwrap().bond_amount, 0);
-    }
-
-    #[test]
-    fn slash_bond_on_missing_agent_returns_not_found() {
-        let (_env, client, _admin) = setup_with_admin();
-        assert_eq!(
-            client.try_slash_bond(&Symbol::new(&_env, "ghost"), &1_000_i128),
-            Err(Ok(Error::NotFound))
-        );
-    }
-
-    #[test]
-    fn slash_bond_requires_admin() {
-        let env = Env::default();
-        let contract_id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &contract_id);
-        let admin = Address::generate(&env);
-        env.mock_all_auths();
-        client.initialize(&admin);
-
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "protected", "research", owner));
-
-        // Non-admin cannot slash.
-        env.mock_auths(&[]);
-        let result = client.try_slash_bond(&Symbol::new(&env, "protected"), &1_000_i128);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn deregister_initiates_cooldown() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "cooldown_agent", "research", owner));
-
-        // First deregister call — should succeed and store a cooldown record.
-        assert!(client
-            .try_deregister_agent(&Symbol::new(&env, "cooldown_agent"))
-            .is_ok());
-
-        // Agent should be gone from the registry immediately.
-        let agents = client.lookup_agents(&Symbol::new(&env, "research"));
-        assert_eq!(agents.len(), 0);
-
-        // But a second call before cooldown elapses should return CooldownNotElapsed.
-        assert_eq!(
-            client.try_deregister_agent(&Symbol::new(&env, "cooldown_agent")),
-            Err(Ok(Error::CooldownNotElapsed))
-        );
-    }
-
-    #[test]
-    fn bond_return_before_cooldown_is_rejected() {
-        let (env, client) = setup();
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "early_return", "research", owner));
-        client.deregister_agent(&Symbol::new(&env, "early_return"));
-
-        // Immediately try to claim the bond — cooldown has not elapsed yet.
-        assert_eq!(
-            client.try_deregister_agent(&Symbol::new(&env, "early_return")),
-            Err(Ok(Error::CooldownNotElapsed))
-        );
-    }
-
-    #[test]
-    fn bond_returned_after_cooldown_elapses() {
-        let env = Env::default();
-        env.mock_all_auths();
-        // Set a very high max_entry_ttl so nothing gets archived when we
-        // advance the ledger past the default min_persistent_entry_ttl.
-        env.ledger().set_max_entry_ttl(100_000_000);
-        env.ledger().set_min_persistent_entry_ttl(100_000_000);
-
-        let id = env.register(AgentRegistryContract, ());
-        let client = AgentRegistryContractClient::new(&env, &id);
-
-        let owner = Address::generate(&env);
-        client.register_agent(&make_record(&env, "wait_agent", "research", owner));
-        client.deregister_agent(&Symbol::new(&env, "wait_agent"));
-
-        // Advance ledger sequence past the 24-hour cooldown window.
-        let new_seq = env.ledger().sequence() + BOND_COOLDOWN_LEDGERS + 1;
-        env.ledger().set_sequence_number(new_seq);
-
-        // Second call after cooldown should succeed and emit BondReturned.
-        assert!(client
-            .try_deregister_agent(&Symbol::new(&env, "wait_agent"))
-            .is_ok());
-
-        // One BondReturned event should have been emitted.
-        let events = env.events().all();
-        assert_eq!(events.len(), 1, "bond return must emit 1 event");
-        assert_event_topics(
-            &env,
-            0,
-            symbol_short!("registry"),
-            symbol_short!("bond_ret"),
-        );
-    }
-
-    #[test]
-    fn estimate_gas_slash_bond_operation() {
-        let (env, client) = setup();
-        let one = client.estimate_gas(&String::from_str(&env, "slash_bond"), &1);
-        let three = client.estimate_gas(&String::from_str(&env, "slash_bond"), &3);
-        assert_eq!(one, GAS_SLASH_BOND);
-        assert_eq!(three, GAS_SLASH_BOND * 3);
-    }
-
-    #[test]
-    fn estimate_gas_deregister_with_bond_operation() {
-        let (env, client) = setup();
-        let one = client.estimate_gas(&String::from_str(&env, "deregister_with_bond"), &1);
-        let two = client.estimate_gas(&String::from_str(&env, "deregister_with_bond"), &2);
-        assert_eq!(one, GAS_DEREGISTER_WITH_BOND);
-        assert_eq!(two, GAS_DEREGISTER_WITH_BOND * 2);
-    }
-}
+mod test;
